# frontend/public/demo-api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaxEvidenceCounts": {
        "description": "How many of a line's transactions carry each kind of evidence.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "integer"
          },
          "receipt": {
            "title": "Receipt",
            "type": "integer"
          },
          "statement": {
            "title": "Statement",
            "type": "integer"
          }
        },
        "required": [
          "receipt",
          "email",
          "statement"
        ],
        "title": "TaxEvidenceCounts",
        "type": "object"
      },
...
      "TaxLineResponse": {
        "description": "One claim line with its matched transactions for the year.",
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "cra_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Tax line reference, rendered verbatim.",
            "title": "Cra Ref"
          },
          "evidence_counts": {
            "$ref": "#/components/schemas/TaxEvidenceCounts"
          },
          "excluded_transactions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaxPackTransaction"
            },
            "title": "Excluded Transactions",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional one-sentence note from the mapping seed.",
            "title": "Note"
          },
          "total": {
            "title": "Total",
            "type": "number"
          },
          "transaction_count": {
            "title": "Transaction Count",
            "type": "integer"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TaxPackTransaction"
            },
            "title": "Transactions",
            "type": "array"
          }
        },
        "required": [
          "key",
          "label",
          "categories",
          "total",
          "transaction_count",
          "evidence_counts",
          "transactions"
        ],
        "title": "TaxLineResponse",
        "type": "object"
      },
...
      "TaxPackResponse": {
        "description": "The full calendar-year tax pack: every mapped line plus a grand total.",
        "properties": {
          "grand_total": {
            "title": "Grand Total",
            "type": "number"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/TaxLineResponse"
            },
            "title": "Lines",
            "type": "array"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "grand_total",
          "lines"
        ],
        "title": "TaxPackResponse",
        "type": "object"
      },
      "TaxPackTransaction": {
        "description": "One claimable transaction inside a tax line.",
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "date": {
            "description": "Transaction date, YYYY-MM-DD.",
            "title": "Date",
            "type": "string"
          },
          "date_file_name": {
            "title": "Date File Name",
            "type": "string"
          },
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "forwarded_to": {
            "title": "Forwarded To",
            "type": "string"
          },
          "manual": {
            "default": false,
            "title": "Manual",
            "type": "boolean"
          },
          "tx_id": {
            "title": "Tx Id",
            "type": "string"
          }
        },
        "required": [
          "tx_id",
          "date",
          "company",
          "amount",
          "category",
          "evidence",
          "forwarded_to",
          "date_file_name"
        ],
        "title": "TaxPackTransaction",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal finance dashboard API. Ingests bank transaction emails, categorizes them (with optional OpenAI enrichment), and exposes transactions, budgets, summaries, and AI spending insights over a consistent JSON HTTP interface.\n\n**Storage:** dual-backend \u2014 DynamoDB (AWS) or SQLite (local/demo). Selection is driven by `data/config.json`. All endpoints return identical shapes regardless of backend.\n\n**Versioning:** routes are prefixed with `/api/v1/`. Breaking changes will ship as `/api/v2/` rather than in-place.\n\n**Demo:** this is a read-only snapshot of the Tidings demo journal \u2014 the fictional Mira Lin Chen household, covering 2025-05 through 2026-03. All data is invented. Any request other than GET, HEAD, or OPTIONS returns 405; self-host Tidings for a writable journal with bearer-token auth.",
    "title": "Tidings demo API (read-only)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tax-pack": {
      "get": {
        "operationId": "getTaxPack",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "maximum": 2099,
              "minimum": 2020,
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxPackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calendar-year tax pack: claim-line totals with per-transaction evidence",
        "tags": [
          "tax"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://gettidings.com/demo"
    }
  ]
}